feat!: Added Basic Report View for Teachers (#15)
* feat: Add Static  Reports Page

* fix: Disable Popup Button Until It's Ready

* BREAKING CHANGE: Updated How Question Answers are Stored in Firestore

This is a precursor to running live reports. Instead of each question having an `answers` sub collection, there is a single `currentQuestionAnswers` sub collection in the class object. Now, when a user joins the class, they get an object in that collection automatically. This means that students who join the class but never answer a single question will be included in the report. Then, once the question is complete, all of those answers are stored as an object directly question. This means that the number of database reads to generate a report goes from (questions * users) per report, to a one time cost of (questions * users), and a per report cost of questions. This results in a significant saving as long as the teacher looks at the report more than once (which is to be expected).

* refactor: Generalized Log Error/Return Function to Support Dynamic Data

* feat: Added Live Report Generation

* fix: Removed Unused Imports in App.tsx