feat: Added Pop Out Window for Classes and Syncing Across Tabs (#21)
* feat: Store Session Info and Listen for Auth State

Now updates all tabs when a user logs in or out, and brings the user back to the classes page if they refresh Auto TA while logged in.

Closes #7

* feat: Added Pop Out Button that Calculates Size Based on User's Screen

Still a work in progress...

* feat: Added Pop Out Window For Classes

Allows the user to more easily rearrange it on the screen.

* fix: Student Question/Answer Options Now Synced Across Tabs

Minor rollback of #13

If the user has multiple tabs open, and answers a question on one tab, all of them should close the submission. While this is not a data integrity issue, as all tabs will close the question when the teacher does (so the student cannot change their answer after the teacher enters the correct answer), it does make for  a better user experience if all the tabs are up to date.